Save point: trades ledger + prices feed, refactor phase 0-2 groundwork
- trades: unified cross-venue sales ledger (dex/dispense/emblem) + /v2/trades read surface + USD price backfill (parallel session)
- packages/shared: wire-contract package scaffold (envelope, models draft)
- apps/api: typed D1 helpers (db.ts); ops/ dir for runbooks + drive scripts; migrate scripts point at xcpio (was xcpdex)
- docs: refactor proposal + architecture pattern language; SESSION-HANDOFF marked resolved
- repo: root workspace scripts, gitignore for logs/token files

+++ apps/web/src/app/assets/page.tsx
@@ -3,7 +3,7 @@ import Link from "next/link";
 import { useState } from "react";
 import { useAssets } from "@/lib/hooks";
 import { Card, Table, Row, Cell, Skeleton, ErrorBox, Empty, AssetIcon, LockBadge, SecondaryButton } from "@/components/ui";
-import { commas, short } from "@/lib/format";
+import { commas } from "@/lib/format";
 
 export default function AssetsPage() {
   const [query, setQuery] = useState("");
@@ -31,7 +31,7 @@ export default function AssetsPage() {
                 </Cell>
                 <Cell muted><span className="block max-w-xs truncate">{a.description || ""}</span></Cell>
                 <Cell numeric>{commas(a.supply_normalized)}</Cell>
-                <Cell muted><Link href={`/address/${a.issuer}`} className="font-mono">{short(a.issuer)}</Link></Cell>
+                <Cell muted><Link href={`/address/${a.issuer}`} className="font-mono">{a.issuer}</Link></Cell>
                 <Cell>{a.locked ? <LockBadge locked /> : <LockBadge />}</Cell>
               </Row>
             ))}
